Basic plugin, more correct example
